+++ src/config/ParseConfig.cpp
@@ -501,20 +501,6 @@ std::string ParseConfig::preprocess(It it, const It end) const {
 
       result.append(begin, it);
     }
-    else if (skip(&it, end, "$(")) {
-      // a terminal command
-      for (auto level = 1; level > 0; ) {
-        if (it == end)
-          error("Unterminated terminal command");
-        if (skip(&it, end, '('))
-          ++level;
-        else if (skip(&it, end, ')'))
-          --level;
-        else
-          ++it;
-      }
-      result.append(begin, it);
-    }
     else {
       // single character
       result.append(begin, ++it);